[060_parent_class_mixed], issue #60, moved some common methods to MixedClass
