+++ dassh/region_mixed.py
@@ -148,20 +148,14 @@ class MixedRegion(RoddedRegion, MixedClass):
                               spacer_grid, byp_ff, byp_k, wwdir, sf, se2, 
                               param_update_tol, rad_isotropic=False)
         # Instantiate MixedClass object
-        MixedClass.__init__(self, self.subchannel.n_sc['coolant']['total'])
+        MixedClass.__init__(self, self.subchannel.n_sc['coolant']['total'],
+                            coolant_mat)
         
-        self._pressure_drop = 0.0 # This overrides the attribute in RoddedRegion
         # Flag to indicate whether to track iteration convergence or not 
         self._verbose = verbose
         # Tolerance for mixed convection solver and star quantities calculation
         self._mixed_convection_rel_tol = mixed_convection_rel_tol
         self._accurate_star_quantities = accurate_star_quantities
-        # Initialize coolant density in subchannels
-        self.sc_properties['density'] = self.coolant.density * \
-            np.ones(self.subchannel.n_sc['coolant']['total']) 
-        # Initialize enthalpy array
-        self._enthalpy = self.coolant.convert_properties(
-            density=self.sc_properties['density'])
 
 
     ####################################################################
@@ -544,9 +538,15 @@ class MixedRegion(RoddedRegion, MixedClass):
         """
         self._calc_h_v_star(delta_v, delta_rho, RR, nn)
         # Calculate coefficients for the matrix
-        EE, FF = self._calc_momentum_coefficients(nn, dz, delta_v)
+        EE, FF = self._calc_momentum_coefficients(
+            nn, dz, self.coolant_int_params['ff_i'], 
+            self.params['de'][self.subchannel.type[:nn]], 
+            delta_v
+            )
         SS, TT = self._calc_energy_coefficients(delta_v, delta_rho, RR)
-        C_rho, C_v = self._calc_continuity_coefficients(nn, delta_v)
+        C_rho, C_v = self._calc_continuity_coefficients(
+            delta_v, self.params['area'][self.subchannel.type[:nn]]
+            )
         # Build matrix
         AA = np.zeros((2*nn + 1, 2*nn + 1))
             
@@ -564,75 +564,6 @@ class MixedRegion(RoddedRegion, MixedClass):
         AA[-1,0:2*nn:2] = C_rho
         AA[-1,1:2*nn:2] = C_v
         return AA
-
-
-    def _calc_h_v_star(self, delta_v: np.ndarray, delta_rho: np.ndarray, 
-                       RR: np.ndarray, nn: int) -> None:
-        """
-        Update hstar and vstar
-        
-        Parameters
-        ----------
-        delta_v : np.ndarray
-            Variation of the SC velocities (m/s)
-        delta_rho : np.ndarray
-            Variation of the SC densities (kg/m^3)
-        RR : np.ndarray
-            Enthalpy variation coefficient (J*m^3/kg^2)
-        nn : int
-            Number of coolant subchannels
-            
-        Notes
-        -----
-        Two options are available:
-        1) Approximate hstar and vstar as the midpoint value of enthalpy
-           and velocity (i.e., at z + dz/2) `h_mid` and `v_mid`
-        2) Calculate hstar and vstar as per "Cheng, S.K., 1984. Constitutive 
-           Correlations for wire-wrapped subchannel analysis under forced and
-           mixed convection conditions (Ph.D. thesis). MIT."
-        """
-        h_mid = self._enthalpy + RR * delta_rho / 2
-        v_mid = self._sc_vel + delta_v / 2
-        # OPTION 1: Approximate hstar and vstar as midpoints
-        if not self._accurate_star_quantities:
-            self._hstar = h_mid
-            self._vstar = v_mid
-            return
-        # OPTION 2: Calculate hstar and vstar as per Cheng 
-        numerator_h = np.zeros(nn)
-        numerator_v = np.zeros(nn)
-        sum_den = np.zeros(nn)
-        # Calculate delta_m for each subchannel
-        vrho_1 = self.sc_properties['density'] * self._sc_vel 
-        vrho_2 = (self.sc_properties['density'] + delta_rho) * \
-            (self._sc_vel + delta_v) 
-        delta_m = (vrho_2 - vrho_1) * \
-            self.params['area'][self.subchannel.type[:nn]]
-        # Iterate over subchannels and adjacent subchannels
-        for i in range(nn):
-            denominator = 0.0
-            num_h = 0.0
-            num_v = 0.0
-            for k in range(3):
-                j = self.ht['cond']['adj'][i][k]
-                if i in self.ht['conv']['ind'][self.ht['conv']['type'] == 2] \
-                    and k == 2:
-                    continue
-                # Calculate delta_m difference between adjacent subchannel
-                xij = delta_m[i] - delta_m[j]
-                # Calculate numerators and denominators
-                num_h += self._calc_star_quantity_numerator(
-                    h_mid[i], h_mid[j], xij)
-                num_v += self._calc_star_quantity_numerator(
-                    v_mid[i], v_mid[j], xij)
-                denominator += np.abs(xij)
-            numerator_h[i] = num_h
-            numerator_v[i] = num_v
-            sum_den[i] = denominator      
-        # Calculate hstar and vstar
-        sum_den = 2 * sum_den + sys.float_info.epsilon 
-        self._hstar = numerator_h / sum_den
-        self._vstar = numerator_v / sum_den
         
         
     def _calc_star_quantity_numerator(self, var_mid_i: float, var_mid_j: float,
@@ -658,97 +589,6 @@ class MixedRegion(RoddedRegion, MixedClass):
         """
         return np.abs(xij) * (var_mid_i + var_mid_j) \
             - xij * (var_mid_i - var_mid_j)
-        
-        
-    def _calc_momentum_coefficients(self, nn: int, dz: float, 
-                                    delta_v: np.ndarray) -> tuple[np.ndarray]:
-        """
-        Calculate Ei and Fi coefficients for the momentum equation
-        
-        Parameters
-        ----------
-        nn : int
-            Number of coolant subchannels
-        dz : float
-            Axial step size (m)
-        delta_v : np.ndarray
-            Variation of the SC velocities (m/s)
-            
-        Returns
-        -------
-        Tuple[np.ndarray]
-            Container of the two following np.ndarrays:
-            
-            - EE coefficients 
-            - FF coefficients
-        """
-        EE = (self._sc_vel + delta_v) * \
-            (self._sc_vel + delta_v - self._vstar) + GRAVITY_CONST * dz / 2 + \
-                self.coolant_int_params['ff_i'] * dz / 16 / \
-                    self.params['de'][self.subchannel.type[:nn]] * \
-                        (2 * self._sc_vel + delta_v)**2 
-        FF = self.sc_properties['density'] * (
-            (2 + self.coolant_int_params['ff_i'] * dz / 2 / 
-             self.params['de'][self.subchannel.type[:nn]]) * self._sc_vel + 
-            (1 + self.coolant_int_params['ff_i'] * dz / 8 /
-             self.params['de'][self.subchannel.type[:nn]]) * delta_v 
-            - self._vstar)
-        return EE, FF
-
-
-    def _calc_energy_coefficients(self, delta_v: np.ndarray, 
-                                  delta_rho: np.ndarray, 
-                                  RR: np.ndarray) -> tuple[np.ndarray]:
-        """
-        Calculate coefficients for the energy equation.
-        
-        Parameters
-        ----------
-        delta_v : np.ndarray
-            Variation of the SC velocities (m/s)
-        delta_rho : np.ndarray
-            Variation of the SC densities (kg/m^3)
-        RR : np.ndarray
-            Enthalpy variation coefficient (J*m^3/kg^2)
-            
-        Returns
-        -------
-        Tuple[np.ndarray]
-            Container of the two following np.ndarrays:
-            
-            - SS coefficients 
-            - TT coefficients
-        """
-        SS = (self._sc_vel + delta_v) * \
-            (self._enthalpy - self._hstar + 
-             RR * (self.sc_properties['density'] + delta_rho))
-        TT = self.sc_properties['density'] * (self._enthalpy - self._hstar)
-        return SS, TT
-
-
-    def _calc_continuity_coefficients(self, nn: int, delta_v: np.ndarray) \
-        -> tuple[np.ndarray]:
-        """
-        Calculate coefficients for the continuity equation.
-        
-        Parameters
-        ----------
-        nn : int
-            Number of coolant subchannels
-        delta_v : np.ndarray
-            Variation of the SC velocities (m/s)
-            
-        Returns
-        -------
-        Tuple[np.ndarray]
-            Container of the two following np.ndarrays:
-            
-            - C_rho coefficients 
-            - C_v coefficients
-        """
-        areas = self.params['area'][self.subchannel.type[:nn]]
-        return areas * (self._sc_vel + delta_v), \
-            areas * self.sc_properties['density']
     
     
     def _calc_RR(self, drho: np.ndarray) -> np.ndarray:
